+++ settings.example-bob.py
@@ -174,16 +174,25 @@ sensors = {
             'enabled': False,
         },
     },
-    'registry': {
-        'hx711': {
+    'environment': [
+        {
+            'key': 'scale-1',
+            'name': 'scale',
+            'description': 'Waage 1',
+            'type': 'HX711',
             'enabled': True,
             'address': 0x00,
             'pin_dout': 'P22',
             'pin_pdsck': 'P21',
             'scale': 4.424242,
-            'offset': -73000.0,
+            'offset': -73000,
         },
-        'ds18x20': {
+        {
+            'key': 'ds18b20-1',
+            'name': 'temperature',
+            'description': 'Wabengasse 1',
+            'type': 'DS18B20',
+            'enabled': True,
             'bus': 'onewire:0',
             'devices': {
                 '1111111111111111': {
@@ -194,14 +203,17 @@ sensors = {
                     'enabled': True,
                     #'offset': -0.42,
                 },
-            }
+            },
         },
-        'bme280': {
+        {
+            'key': 'bme280-1',
+            'description': 'Temperatur und Feuchte außen',
+            'type': 'BME280',
             'enabled': True,
             'bus': 'i2c:0',
             'address': 0x77,
-        },
-    },
+        }
+    ],
     'busses': [
         {
             "family": "i2c",